+++ packages/core/src/components/Checkbox/Checkbox.stories.tsx
@@ -1,24 +1,29 @@
 import { CheckboxTheme, defaultTheme } from '@medly-components/theme';
-import { boolean, select } from '@storybook/addon-knobs';
-import React from 'react';
+import React, { useState } from 'react';
+import Button from '../Button';
 import { Checkbox } from './Checkbox';
 import { Props } from './types';
-
-const labelPosition: Props['labelPosition'][] = ['left', 'right', 'top', 'bottom'];
-const size: Props['size'][] = ['XS', 'S', 'M', 'L', 'XL'];
+export const labelPositions: Props['labelPosition'][] = ['left', 'right', 'top', 'bottom'];
+export const sizes: Props['size'][] = ['XS', 'S', 'M', 'L', 'XL'];
 
 export const ThemeInterface: React.FC<CheckboxTheme> = () => null;
 ThemeInterface.defaultProps = {
     ...defaultTheme.checkbox
 };
 
-export const Basic = () => (
-    <Checkbox
-        name="fruit"
-        size={select('Size', size, 'S')}
-        label="Orange"
-        fullWidth={boolean('Full Width', false)}
-        disabled={boolean('Disabled', false)}
-        labelPosition={select('Label Position', labelPosition, 'right')}
-    />
-);
+export const FormWithCheckbox: React.FC = props => {
+    const [checked, setChecked] = useState<boolean>(),
+        onChangeHandler = (e: React.ChangeEvent<HTMLInputElement>) => setChecked(e.target.checked),
+        handleSubmit = (e: React.FormEvent) => e.preventDefault();
+    return (
+        <form onSubmit={handleSubmit}>
+            <Checkbox fullWidth required checked={checked} onChange={onChangeHandler} {...props} />
+            <Button size="S" type="submit">
+                Submit
+            </Button>
+        </form>
+    );
+};
+FormWithCheckbox.defaultProps = {
+    label: 'By clicking Submit, I agree lorem ipsum'
+};