+++ packages/core/src/components/Checkbox/checkbox.test.tsx
@@ -1,24 +1,65 @@
-import { fireEvent, render } from '@test-utils';
+import { fireEvent, render, screen } from '@test-utils';
+import userEvent from '@testing-library/user-event';
 import React from 'react';
+import { LabelPositions } from '../Label/types';
 import { Checkbox } from './Checkbox';
 
 describe('Checkbox component', () => {
-    it('should render correctly with all the default props', () => {
-        const { container } = render(<Checkbox name="gender" />);
+    describe.each([true, false])('should render correctly with checked %p', (checked: boolean) => {
+        test.each(['indeterminate', 'active', 'disabled', 'hasError'])('and %s state', state => {
+            const { container } = render(<Checkbox {...{ [state]: true }} checked={checked} errorText="Something went wrong" />);
+            expect(container).toMatchSnapshot();
+        });
+    });
+
+    test.each(['left', 'right', 'top', 'bottom'])('should render correctly with %s labelPosition', (labelPosition: LabelPositions) => {
+        const { container } = render(<Checkbox label="Dummy" labelPosition={labelPosition} />);
         expect(container).toMatchSnapshot();
     });
 
-    it('should render correctly with all the props given', () => {
-        const { container } = render(<Checkbox disabled fullWidth name="gender" label="Female" size="M" labelPosition="right" />);
-        expect(container).toMatchSnapshot();
+    test.each(['top', 'bottom'])('should render properly with fullWidth and %p labelPosition', (labelPosition: LabelPositions) => {
+        const { container } = render(<Checkbox id="dummy" fullWidth label="Dummy" labelPosition={labelPosition} />);
+        expect(container.querySelector('#dummy-wrapper')).toHaveStyle(`
+            display: flex;
+            padding: 0.8rem 0
+        `);
     });
 
     it('should call onChange handler on click on checkbox', () => {
         const mockOnChange = jest.fn();
-        const { container } = render(
-            <Checkbox onChange={mockOnChange} disabled name="gender" label="Female" size="M" labelPosition="right" />
-        );
-        fireEvent.click(container.querySelector('input'));
+        render(<Checkbox onChange={mockOnChange} label="Dummy" />);
+        userEvent.click(screen.getByRole('checkbox'));
         expect(mockOnChange).toBeCalled();
+    });
+
+    it('should show error text if provided', () => {
+        const mockOnChange = jest.fn();
+        render(<Checkbox onChange={mockOnChange} label="Dummy" errorText="Something went wrong" />);
+        const errorText = screen.getByText('Something went wrong');
+        expect(errorText).toBeInTheDocument();
+        expect(errorText).toHaveStyle(`color: rgb(204, 0, 0)`);
+    });
+
+    it('should call validator function on blur event', () => {
+        const mockOnBlur = jest.fn();
+        render(
+            <Checkbox
+                required
+                onBlur={mockOnBlur}
+                label="Dummy"
+                validator={val => (!val ? 'Please accept the terms & conditions to continue.' : '')}
+            />
+        );
+        fireEvent.blur(screen.getByRole('checkbox'));
+        expect(screen.getByText('Please accept the terms & conditions to continue.')).toBeInTheDocument();
+        expect(mockOnBlur).toBeCalled();
+    });
+
+    it('should show the html5 error validation message on error', () => {
+        const mockOnInvalid = jest.fn();
+        render(<Checkbox required onInvalid={mockOnInvalid} label="Dummy" />);
+        fireEvent.invalid(screen.getByRole('checkbox'));
+        expect(screen.getByText('Constraints not satisfied')).toBeInTheDocument();
+        expect(mockOnInvalid).toBeCalled();
     });
 });