+++ packages/core/src/components/CheckboxGroup/CheckboxGroup.tsx
@@ -1,29 +1,66 @@
-import { WithStyle } from '@medly-components/utils';
-import React, { FC, useCallback, useMemo } from 'react';
+import { useCombinedRefs, WithStyle } from '@medly-components/utils';
+import React, { FC, useCallback, useMemo, useState } from 'react';
 import Checkbox from '../Checkbox';
-import FieldWithLabel from '../FieldWithLabel';
+import { SelectorGroup } from '../Selectors';
 import getValuesFromOptions from './getValuesFromOptions';
 import { Props } from './types';
 
 export const CheckboxGroup: FC<Props> & WithStyle = React.memo(
     React.forwardRef((props, ref) => {
         const {
+            id,
             values,
             onChange,
             options,
             size,
             label,
-            labelPosition,
             labelVariant,
             labelWeight,
-            labelColor,
             showSelectAll,
             disabled,
-            fullWidth
+            errorText,
+            helperText,
+            columns,
+            required,
+            validator,
+            parentHasError,
+            ...wrapperProps
         } = props;
 
-        const allChildValues = useMemo(() => getValuesFromOptions(options), [options]),
-            areAllValuesSelected = useMemo(() => allChildValues.every(val => values.includes(val)), [options, values]);
+        const [builtInErrorMessage, setErrorMessage] = useState(''),
+            checkboxGroupId = useMemo(() => id || label, [id, label]),
+            checkboxGroupRef = useCombinedRefs<HTMLDivElement>(ref, React.useRef(null)),
+            hasError = useMemo(() => !!errorText || !!builtInErrorMessage || parentHasError, [
+                builtInErrorMessage,
+                errorText,
+                parentHasError
+            ]),
+            hasHelperOrErrorText = useMemo(() => !!(errorText || helperText), [errorText, helperText]),
+            allChildValues = useMemo(() => getValuesFromOptions(options), [options]),
+            areAllValuesSelected = useMemo(() => allChildValues.every(val => values.includes(val)), [options, values]),
+            indeterminate = useMemo(() => !areAllValuesSelected && allChildValues.some(el => values.includes(el)), [
+                areAllValuesSelected,
+                allChildValues
+            ]);
+
+        const validate = useCallback(
+                selectedValues => {
+                    const message =
+                        (validator && validator(selectedValues)) ||
+                        (required && selectedValues.length === 0 && 'Please select at least one option.') ||
+                        '';
+                    setErrorMessage(message);
+                },
+                [validator]
+            ),
+            onBlur = useCallback(
+                (event: React.FocusEvent<HTMLDivElement>) => {
+                    const currentTarget = event.currentTarget;
+                    setTimeout(() => !currentTarget.contains(document.activeElement) && validate(values), 0);
+                    props.onBlur && props.onBlur(event);
+                },
+                [validate, values, props.onBlur]
+            );
 
         const handleOptionClick = useCallback(
                 (event: React.ChangeEvent<HTMLInputElement>) => {
@@ -31,6 +68,7 @@ export const CheckboxGroup: FC<Props> & WithStyle = React.memo(
                     const item = event.target.name,
                         isChecked = event.target.checked,
                         newValues = isChecked ? [...values, item] : values.filter(vl => vl !== item);
+                    validate(newValues);
                     onChange(newValues);
                 },
                 [values, onChange]
@@ -39,35 +77,68 @@ export const CheckboxGroup: FC<Props> & WithStyle = React.memo(
                 const newValues = areAllValuesSelected
                     ? values.filter(val => !allChildValues.includes(val))
                     : Array.from(new Set([...values, ...allChildValues]));
+                validate(newValues);
                 onChange(newValues);
             }, [options, values, onChange]);
 
         return (
-            <FieldWithLabel id={`${label}-checkboxGroup`} ref={ref} fullWidth {...{ fullWidth, labelPosition }}>
-                {label && (
-                    <FieldWithLabel.Label {...{ labelPosition, labelVariant, labelWeight, labelColor }}>
-                        {showSelectAll ? (
-                            <Checkbox
-                                key="select-all"
-                                {...{ size, disabled, labelVariant, labelWeight, labelColor }}
-                                label={label}
-                                checked={areAllValuesSelected}
-                                onChange={handleSelectAllClick}
-                            />
-                        ) : (
-                            `${label}`
-                        )}
-                    </FieldWithLabel.Label>
+            <SelectorGroup.Wrapper
+                ref={checkboxGroupRef}
+                type="checkbox"
+                id={`${checkboxGroupId}-wrapper`}
+                aria-describedby={`${checkboxGroupId}-helper-text`}
+                {...{ ...wrapperProps, hasHelperOrErrorText, onBlur }}
+            >
+                {label &&
+                    (showSelectAll ? (
+                        <Checkbox
+                            key="select-all"
+                            label={label}
+                            checked={areAllValuesSelected}
+                            onChange={handleSelectAllClick}
+                            indeterminate={indeterminate}
+                            id={`${checkboxGroupId}-select-all`}
+                            {...{ size, hasError, disabled, labelVariant, labelWeight }}
+                        />
+                    ) : (
+                        <SelectorGroup.Label
+                            id={`${checkboxGroupId}-label`}
+                            type="checkbox"
+                            hasError={hasError}
+                            required={required}
+                            disabled={disabled}
+                            textVariant={labelVariant}
+                            textWeight={labelWeight}
+                        >
+                            {label}
+                        </SelectorGroup.Label>
+                    ))}
+                {(errorText || builtInErrorMessage || helperText) && (
+                    <SelectorGroup.HelperText
+                        id={`${checkboxGroupId}-helper-text`}
+                        type="checkbox"
+                        disabled={disabled}
+                        hasError={hasError}
+                        isIndented={showSelectAll}
+                    >
+                        {errorText || builtInErrorMessage || helperText}
+                    </SelectorGroup.HelperText>
                 )}
-                <FieldWithLabel.Field isIndented={labelPosition === 'top'}>
+                <SelectorGroup.Options id={`${checkboxGroupId}-options`} columns={columns} isIndented={showSelectAll}>
                     {options.map(option => {
                         return Array.isArray(option.value) ? (
                             <CheckboxGroup
                                 key={option.label}
+                                parentHasError={hasError}
+                                id={`${option.label}-${checkboxGroupId}`}
                                 {...{
                                     ...props,
+                                    showSelectAll: option.showSelectAll ?? true,
                                     options: option.value,
-                                    label: option.label
+                                    label: option.label,
+                                    columns: option.columns,
+                                    helperText: option.helperText,
+                                    errorText: option.errorText
                                 }}
                             />
                         ) : (
@@ -79,21 +150,22 @@ export const CheckboxGroup: FC<Props> & WithStyle = React.memo(
                                 checked={values.includes(option.value)}
                                 onChange={handleOptionClick}
                                 disabled={disabled || option.disabled}
+                                hasError={hasError}
+                                id={`${option.label}-${checkboxGroupId}`}
                             />
                         );
                     })}
-                </FieldWithLabel.Field>
-            </FieldWithLabel>
+                </SelectorGroup.Options>
+            </SelectorGroup.Wrapper>
         );
     })
 );
 
 CheckboxGroup.displayName = 'CheckboxGroup';
-CheckboxGroup.Style = FieldWithLabel.Style;
+CheckboxGroup.Style = SelectorGroup.Wrapper;
 CheckboxGroup.defaultProps = {
-    disabled: false,
-    fullWidth: true,
     values: [],
-    showSelectAll: false,
-    labelPosition: 'left'
+    columns: 1,
+    labelVariant: 'body1',
+    labelWeight: 'Medium'
 };