+++ packages/core/src/components/RadioGroup/RadioGroup.test.tsx
@@ -1,47 +1,64 @@
-import { cleanup, fireEvent, render } from '@test-utils';
+import { fireEvent, render, screen, waitFor } from '@test-utils';
+import userEvent from '@testing-library/user-event';
 import React from 'react';
 import { RadioGroup } from './RadioGroup';
 
-describe('Radio Group', () => {
-    afterEach(cleanup);
+const renderer = ({
+    label = 'Gender',
+    name = 'gender',
+    helperText = 'Helper Text',
+    errorText = undefined,
+    onChange = jest.fn(),
+    required = false,
+    validator = undefined,
+    onInvalid = jest.fn(),
+    onBlur = jest.fn(),
+    disabled = false,
+    options = [
+        { value: 'female', label: 'Female' },
+        { value: 'male', label: 'Male' }
+    ]
+}) =>
+    render(<RadioGroup {...{ required, disabled, label, name, onInvalid, onBlur, validator, helperText, options, errorText, onChange }} />);
 
-    it('should render correctly', () => {
-        const { container } = render(
-            <RadioGroup
-                disabled
-                fullWidth
-                label="Gender"
-                name="gender"
-                value="female"
-                options={[
-                    { value: 'female', label: 'Female' },
-                    { value: 'male', label: 'Male' }
-                ]}
-            />
-        );
+describe('Radio Group', () => {
+    it('should render properly', () => {
+        const { container } = renderer({});
         expect(container).toMatchSnapshot();
+    });
+
+    it('should render disabled state properly', () => {
+        const { container } = renderer({ disabled: true });
+        expect(container).toMatchSnapshot();
+    });
+
+    it('should be able to render error text', () => {
+        renderer({ errorText: 'Something went wrong' });
+        const errorText = screen.getByText('Something went wrong');
+        expect(errorText).toBeInTheDocument();
+        expect(errorText).toHaveStyle(`color: rgb(204, 0, 0)`);
     });
 
     it('should call onChange prop with selected option', () => {
         const mockOnChange = jest.fn();
-
-        const { container, getByText } = render(
-            <RadioGroup
-                label="Gender"
-                name="gender"
-                onChange={mockOnChange}
-                value="female"
-                required
-                size="M"
-                labelPosition="top"
-                options={[
-                    { value: 'female', label: 'Female' },
-                    { value: 'male', label: 'Male' }
-                ]}
-            />
-        );
-        fireEvent.click(getByText('Male'));
+        renderer({ onChange: mockOnChange });
+        userEvent.click(screen.getByText('Male'));
         expect(mockOnChange).toBeCalledWith('male');
-        expect(container).toMatchSnapshot();
+    });
+
+    it('should show the html5 form validation on error', () => {
+        renderer({ required: true });
+        fireEvent.invalid(screen.getByRole('radio', { name: 'Female' }));
+        expect(screen.getByText('Constraints not satisfied')).toBeInTheDocument();
+    });
+
+    it('should call validator function on blur', async () => {
+        const mockOnBlur = jest.fn();
+        renderer({
+            onBlur: mockOnBlur,
+            validator: () => 'Please select any one gender'
+        });
+        fireEvent.blur(screen.getByRole('radio', { name: 'Female' }));
+        await waitFor(() => expect(screen.getByText('Please select any one gender')).toBeInTheDocument());
     });
 });