+++ packages/utils/src/style/index.ts
@@ -1,3 +1,4 @@
+export * from './asterisk';
 export * from './centerAligned';
 export * from './clearMarginPadding';
 export * from './fullHeight';